Prefab for actionText and organizing Cen02 scripts
Now actionText is a prefab that can be used in any scene.

On Cen02_Cozinha only one script was controlling everything, now there
are two separated scripts: one for the Fridge and another one to leave
the scene.

We are not using as many "Find" as before, trying to get rid of all of
them!