Prefabs everywhere and separated scripts
The camera now is a prefab and is easier to set the point where it moves
to the left to the right, also you can set all of it already on the
editor (piece of cake!)

The disposition bar is inside the Canvas now and it's a prefab so it can
be resized without any bugs anymore.

All events now have their own script and they're prefabs for easy
management and instantiation, tried my best to make sure that any
customizable settings is available directly on Unity editor.

Better organization for folders :smile: hope this can help!